requirements.txt
fastapi>=0.115.0
fastmcp>=2.11.2
starlette>=0.37.2,<0.39.0
rapidfuzz>=3.10.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.32.0
pytest>=8.3.0
pytest-cov>=6.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
ruff>=0.12.0
